NVIDIA-NeMo/requirements.txt
huggingface_hub
numba
numpy>=1.22
onnx>=1.7.0
python-dateutil
ruamel.yaml
scikit-learn
setuptools==59.5.0
tensorboard
text-unidecode
torch
tqdm>=4.41.0
wget
wrapt
braceexpand
editdistance
g2p_en
inflect
ipywidgets
jiwer
kaldi-python-io
kaldiio
librosa>=0.9.0
marshmallow
matplotlib
packaging
pyannote.core
pyannote.metrics
pydub
ruamel.yaml
scipy>=0.14
soundfile
sox
texterrors
pandas
sacremoses>=0.0.43
sentencepiece<1.0.0
youtokentome>=1.0.5
hydra-core>=1.2.0,<1.3
omegaconf>=2.2,<2.3
pytorch-lightning>=1.9.0
pyyaml<6  # Pinned until omegaconf works with pyyaml>=6
#torchmetrics>=0.11.0
transformers>=4.0.1
wandb
webdataset>=0.1.48,<=0.1.62
torchmetrics==0.10.3